requirements.txt
Django~=5.1.5
djangorestframework~=3.15.2
phonenumbers~=8.13.52
twilio~=9.4.3